Remove MAX_NUMCORE.
This fixes a bug I hit in TxQuantize.cpp: the params array is
initialized with length MAX_NUMCORE, but accessed at positions up to
numcore-1. The numcore variable is not constrained to MAX_NUMCORE, so on
systems with cores greater than MAX_NUMCORE, there can be a segfault.

Is there any downside? I don't see one; I'm not familiar with the code,
but it looks like this may cause a negligible decrease in memory usage
on systems with less than MAX_NUMCORE (8) cores, and negligibly higher
memory usage on systems with a higher number of cores as the price of
not segfaulting.

I do not know if I have exercised all code paths in testing this change,
but the usage of MAX_NUMCORE seems the same in all cases.