fix presumed off-by-one bug in TxQuantize::quantize
The surrounding code is structured as if it expects the loop to run as
'< numcore'. Changing the loop conditional makes it consistent with
all the other 'numcore' loops in the file, and is presumably the correct
fix. Unfortunately, I do not know how to exercise this piece of code, so
the patch is given as-is.

Discussed in:
https://github.com/mupen64plus/mupen64plus-video-glide64mk2/issues/100