rdp: Avoid signed overflow during texture block loading
It isn't well defined how the compiler should behave when a signed integer
overlows. This can lead to artifacts like scrambled textures when such an
overflow happens.

For the input v16 == 1879048192 (0x70000000) and dxt == 268435456
(0x10000000) , it would calculate the result 536870912 (0x20000000) instead
of -1 (2147483648 aka 0x80000000). Which means that the flipping will not
be performed correctly.

Fixes: mupen64plus/mupen64plus-video-glide64mk2#120
Signed-off-by: Sven Eckelmann <sven@narfation.org>